Add doc checker and fix some references
Our documentation had a lot of broken links, but it doesn't make the build fail,
so anyone could keep merging with a broken documentation. Now we make the build to fail in case we found broken links in the docs.

Fix broken links and add some tips on how to reference other functions and classes.

ASIM-4127