Add log to alfasim_run when simulation fails
I added this exception handler in order to have more detailed info
when the simulation fails. Before this change, the execution only
returned the subprocess exit code.

ASIM-3405

Change pre-commit's python version to 3.10

Add test of invalid results for the alfasim_runner

Expand coverage for the new feature that prints the logs of
the invalid results in case of the simulation failure due to
exit code != 0

Include log_calc.txt and raise a RuntimeError instead of printing

I've decided to raise a RuntimeError with the logs when the code gets a
CalledProcessorError, because in such manner I am able to catch this
exception in the test

Furthermore, I have removed the check if Results is not None,
because the Results is a class and it could never be None.

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Check if log files exist before raising exception

Co-authored-by: Martin Prüsse <prusse.martin@gmail.com>

Check content of log file in the test

Co-authored-by: Martin Prüsse <prusse.martin@gmail.com>

Load base to alfasim runner directly from alfacase

Co-authored-by: Martin Prüsse <prusse.martin@gmail.com>

Update log message generation

Co-authored-by: Martin Prüsse <prusse.martin@gmail.com>

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

Add coverage ignore comment

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci