Add rod pump support
Add support for the sucker rod pump, also known as just Rod Pump (RP).

New inputs:

plunger_diameter: the transverse size of the plunger responsible to move
the liquid inside the rod pump

stroke_length: the displacement size that the pump is linearly moved up
and down.

pump_speed: the frequency in which the pump moves linearly moved up and
down

slip_coefficient: a coefficient that relates the pump pressure
difference and the slippage flow rate

volumetric_efficiency: a parameter to account the imperfections of the
pump when transporting the liquid.

ASIM-6432